+++ src/components/index.js
@@ -23,7 +23,7 @@ import DpEditor from './DpEditor'
 import DpEditableList from './DpEditableList'
 import DpFlyout from './DpFlyout'
 import DpFormRow from './DpFormRow'
-import { getFileIdsByHash } from './DpUpload/utils/GetFileIdsByHash'
+import { getFileIdsByHash } from './DpUploadFiles/utils/GetFileIdsByHash'
 import DpIcon from './DpIcon'
 import DpInlineNotification from './DpInlineNotification'
 import DpInput from './DpInput'
@@ -55,7 +55,7 @@ import DpTooltip from './DpTooltip'
 import DpTooltipIcon from './DpTooltipIcon'
 import DpTransitionExpand from './DpTransitionExpand'
 import DpTreeList from './DpTreeList'
-import DpUploadFiles from './DpUpload/DpUploadFiles'
+import DpUploadFiles from './DpUploadFiles/DpUploadFiles'
 import DpVideoPlayer from './DpVideoPlayer'
 
 export {
